Student Charges page
- Stylesheet (prototype) updated for the student_charges page
- student_charges (prototype) added
- I believe <div id='content'> should not be in the template. It is not
needed on every page. So removed from PageTemplate and added to
CreateInstructorServlet and my StudentChargesServlet
- PageTemplate's navbar is the administrator's so I added a Student
Charges link to it
- web.xml I added Student Charges mapping
- Stylesheet updated for my StudentCharges page, also extended the
min-width because navbar broke at smaller widths with Student Charges
link
- index.html added link to Student Charges
- Added StudentChargesServlet.java which currently does not properly get
data from database. If data is hardcoded in for names, emails,
classlist, and charges, the html works fine and puts the data in the
table.
- Don't know why it says I added InfiniteLoopServelet.java, I didn't, I
don't think. So disregard.